+++ simintr.h
@@ -16,8 +16,8 @@ class karte_ansicht_t;
 
 bool reduce_frame_time();
 bool increase_frame_time();
-long get_frame_time();
-void set_frame_time(long time);
+uint32 get_frame_time();
+void set_frame_time(uint32 time);
 
 
 void intr_refresh_display(bool dirty);
@@ -29,8 +29,8 @@ void intr_set(karte_t *welt, karte_ansicht_t *view);
  * currently only used by the pause tool. Use with care!
  * @author Hj. Malthaner
  */
-void intr_set_last_time(long time);
-long intr_get_last_time();
+void intr_set_last_time(uint32 time);
+uint32 intr_get_last_time();
 
 
 void intr_enable();
