+++ simsys_d.cc
@@ -423,7 +423,7 @@ void ex_ord_update_mx_my()
 }
 
 
-static volatile long milli_counter;
+static volatile uint32 milli_counter;
 
 void timer_callback()
 {
@@ -454,7 +454,7 @@ static void simtimer_init()
 }
 
 
-unsigned long dr_time()
+uint32 dr_time()
 {
 	return milli_counter;
 }
