+++ vehicle/simverkehr.h
@@ -115,7 +115,7 @@ public:
 
 	const stadtauto_besch_t *get_besch() const { return besch; }
 
-	bool sync_step(long delta_t);
+	bool sync_step(uint32 delta_t);
 
 	grund_t* hop();
 	bool ist_weg_frei(grund_t *gr);
